move open-ai stubbed in UI spec to it's own module
