+++ code/ext/ext.lib.automerge.webrtc/src/test.ui/entry.Specs.mts
@@ -10,7 +10,6 @@ export const Specs = {
   [`${ns}.ui.PeerRepoList`]: () => import('../ui/ui.PeerRepoList/-SPEC'),
   [`${ns}.ui.Sample.01`]: () => import('../ui/ui.Sample.01/-SPEC'),
   [`${ns}.ui.Sample.02`]: () => import('../ui/ui.Sample.02/-SPEC'),
-  [`${ns}.ui.Sample.api.openai`]: () => import('../ui/ui.Sample.AI/-SPEC'),
 };
 
 export default Specs;
