+++ code/ext/ext.lib.automerge.webrtc/src/ui/Icons.ts
@@ -1,11 +1,4 @@
-import {
-  MdArrowUpward,
-  MdCheck,
-  MdClose,
-  MdRefresh,
-  MdSend,
-  MdSettingsInputAntenna,
-} from 'react-icons/md';
+import { MdArrowUpward, MdClose, MdRefresh, MdSettingsInputAntenna } from 'react-icons/md';
 import { TbDatabase } from 'react-icons/tb';
 import { VscSymbolClass } from 'react-icons/vsc';
 import { Icon } from 'sys.ui.react.common';
@@ -22,6 +15,4 @@ export const Icons = {
   Database: icon(TbDatabase),
   Antenna: icon(MdSettingsInputAntenna),
   Object: icon(VscSymbolClass),
-  Send: icon(MdSend),
-  Check: icon(MdCheck),
 } as const;
