+++ code/ext/ext.lib.openai/package.json
@@ -13,6 +13,8 @@
   "dependencies": {
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "ext.lib.monaco": "0.0.0",
+    "ext.lib.monaco.crdt": "0.0.0",
     "sys.util": "0.0.0",
     "sys.ui.react.common": "0.0.0"
   },
