+++ code/ext/ext.lib.openai/src/index.mts
@@ -11,6 +11,9 @@ export { Pkg };
 /**
  * Dev
  */
+import SpecSample from './ui/ui.Sample.01/-SPEC';
+export const Sample = { spec: SpecSample } as const;
+
 export const dev = async () => {
   const { Specs } = await import('./test.ui/entry.Specs.mjs');
   return { Pkg, Specs };
