+++ code/ext/ext.lib.openai/src/ui/common/Icons.ts
@@ -1,4 +1,4 @@
-import { MdRefresh } from 'react-icons/md';
+import { MdCheck, MdRefresh, MdSend } from 'react-icons/md';
 import { Icon } from 'sys.ui.react.common';
 
 const icon = Icon.renderer;
@@ -8,4 +8,6 @@ const icon = Icon.renderer;
  */
 export const Icons = {
   Refresh: icon(MdRefresh),
+  Check: icon(MdCheck),
+  Send: icon(MdSend),
 } as const;
