+++ code/ext/ext.lib.openai/src/ui/ui.Sample.01/ui.tsx
